+++ src/internal/sorters/by-string.ts
@@ -1,0 +1,11 @@
+export function sortByStringAsc(a: string, b: string): number {
+    if (a < b) { return -1; }
+    if (a > b) { return 1; }
+    return 0;
+}
+
+export function sortByStringDesc(a: string, b: string): number {
+    if (a < b) { return 1; }
+    if (a > b) { return -1; }
+    return 0;
+}