+++ src/model/sorter.ts
@@ -1,0 +1,3 @@
+export interface Sorter<T> {
+    sort: (a: T, b: T) => number
+}