+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ### 0.2.5
+- Added sort option with possibility to add custom sorter
 - Added unique option to filter unique array values
 - Added changelog.md file to track changes